% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bollen_plot.R
\name{bollen_plot}
\alias{bollen_plot}
\title{Bollen Influential Cases Plot}
\usage{
bollen_plot(x, crit.value = NULL)
}
\arguments{
\item{x}{A dataset or matrix}

\item{crit.value}{Should a critical value be used to identify influential row names?}
}
\description{
This codes implements the Bollen influential cases plot described in Bollen (1989). This function is unable to handle missing data and listwise deletion is used. On the y-axis is the measure of influence, A, see Bollen (1989) for derivation,  and the row names are on the x-axis. These row names correspond to the original data set's row names prior to listwise deletion.
}
\examples{
data(interest)
bollen_plot(interest)

}
\references{
Bollen, K. A. (1989). Structural equations with latent variables. John Wiley & Sons.
}

