% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_alpha.R
\name{coef_alpha}
\alias{coef_alpha}
\title{Coefficient alpha}
\usage{
coef_alpha(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
Coefficient alpha.
}
\description{
Calculates coefficient alpha, i.e., Cronbach's alpha
}
\examples{
data(SAPA)
coef_alpha(SAPA)
}
