% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvr.R
\name{cvr}
\alias{cvr}
\title{CVR}
\usage{
cvr(N, n_e)
}
\arguments{
\item{N}{total number of experts}

\item{n_e}{number of experts that agree that the item is essential.}
}
\value{
CVR.
}
\description{
Calculates CVR
}
\examples{
cvr(20, 17)
}
