% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{depression}
\alias{depression}
\title{Fabricated Multidimensional Depression Scale}
\format{A polytomous response data set with 2000 examinees and 20 items.
\describe{The item names range from "item1" to "item20". All the items are polytomously scored with five response categories (i.e.,0=Strongly disagree;
1=Disagree; 2=Neither agree or disagree; 3=Agree; and 4=Strongly agree). There is no missing data.
}}
\usage{
depression
}
\description{
This fabricated data set presents responses to a hypothetical depression scale. The scale consists of 20 items, where the first
10 items measure the cognitive symptoms of depression, while the second set of 10 items measure the somatic symptoms of depression.
The two dimensions are moderately correlated (i.e., around .50).
}
\keyword{dataset}
