% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distract.R
\name{distract}
\alias{distract}
\title{distract}
\usage{
distract(data)
}
\arguments{
\item{data}{data frame}
}
\value{
The proportion endorsing a distractor.
}
\description{
Calculates the proportion endorsing a distractor.
}
\examples{
distract(multiplechoice)
}
