% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstudy.R
\name{dstudy}
\alias{dstudy}
\title{Conduct a dependability study}
\usage{
dstudy(x, n, unit)
}
\arguments{
\item{x}{An lmer fitted object}

\item{n}{A vector with facets named}

\item{unit}{The label for the unit of measurement}
}
\description{
Generates the variance components and the relative magnitude (percent of total variance) of each variance component.
}
\examples{
## One facet design
one.facet <- lmer(Score ~ (1 | Participants) + (1 | Items), data = efData)
dstudy(one.facet, n = c("Items" = 4), unit = "Participants")

## Two facet design
lmer.tab3.2 <- lmer(score ~ (1 | child) + (1 | raters) + (1 | occasions) + (1 | child:raters) + (1 | child:occasions) + (1 | raters:occasions), data = swtab1.3)
dstudy(lmer.tab3.2, unit = "child", n = c("occasions" = 4, "raters" = 2))
}

