% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{efData}
\alias{efData}
\title{Hypothetical Executive Functioning Assessment}
\format{A data frame in long-format with 30 participants' responses to 10 items.
 \describe{
\item{participant}{The participant ID.}
\item{item}{The item ID.}
\item{score}{A participant's score on a particular item.}
}}
\usage{
efData
}
\description{
The executive function assessement consists of 10 items measured on 30 student participants. The data were randomly generated to demonstrate various psychometric models (e.g., CTT, a one-facet G-study)
}
\keyword{dataset}

