% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fadiag_demo.R
\name{fadiag_demo}
\alias{fadiag_demo}
\title{Diagnostic plot for factor analysis Shiny demo}
\usage{
fadiag_demo()
}
\description{
This code is a demonstration of Shiny and how Shiny can be used to explore violations to the factor analytic model.
}
\examples{
fadiag_demo()
}
