% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/free_params.R
\name{free_params}
\alias{free_params}
\title{Calculate the number of estimated parameters}
\usage{
free_params(x)
}
\arguments{
\item{x}{A lavaan fitted object}
}
\description{
Given a fitted objects from lavaan, this functions returns the number of free parameter. Note, this function will give incorrect results when intercepts are estimated and should be not used.
}
\examples{
iq.cfa <- '
verb =~ info + comp + arith + simil + digit + vocab
perf =~ pictcomp + parang + block + object + coding
'
fit <- cfa(iq.cfa, data = wiscsem)
free_params(fit)
}

