% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gstudy.R
\name{gstudy}
\alias{gstudy}
\title{Generate the variance components from a G-study}
\usage{
gstudy(x)
}
\arguments{
\item{x}{An lmer fitted object}
}
\description{
Generates the variance components and the relative magnitude (percent of total variance) of each variance component.
}
\examples{
mod <- lmer(score ~ (1 | participant) + (1 | item), data = efData)
gstudy(mod)
}

