% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hcre}
\alias{hcre}
\title{Hypothetical College Readiness Exam}
\format{A 1000 by 2 data.frame containg a student's score and the form of the exam.
 \describe{
\item{score}{The student's score on the college readiness exam}
\item{form}{The form of the exam a student took, either x or y}
}}
\usage{
hcre
}
\description{
This datasets contains score distributions for two forms of a hypothetical college readiness exam.
}
\keyword{dataset}

