% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HSQ}
\alias{HSQ}
\title{Humor Styles Questionnaire}
\format{A data frame with 1070 participants who answered 32 items about their humor styles. The variables Q1 through Q32 were statements rated on a five point scale where 1=Never or very rarely true, 2=Rarely true, 3= Sometimes true, 4= Often true, 5=Very often or always true. The following are the items in HSQ:
\describe{
\item{Q1}{I usually don't laugh or joke around much with other people.}
\item{Q2}{If I am feeling depressed, I can usually cheer myself up with humor.}
\item{Q3}{If someone makes a mistake, I will often tease them about it.}
\item{Q4}{I let people laugh at me or make fun at my expense more than I should.}
\item{Q5}{I don't have to work very hard at making other people laugh-I seem to be a naturally humorous person.}
\item{Q6}{Even when I'm by myself, I'm often amused by the absurdities of life.}
\item{Q7}{People are never offended or hurt by my sense of humor.}
\item{Q8}{I will often get carried away in putting myself down if it makes my family or friends laugh.}
\item{Q9}{I rarely make other people laugh by telling funny stories about myself.}
\item{Q10}{If I am feeling upset or unhappy I usually try to think of something funny about the situation to make myself feel better.}
\item{Q11}{When telling jokes or saying funny things, I am usually not very concerned about how other people are taking it.}
\item{Q12}{I often try to make people like or accept me more by saying something funny about my own weaknesses, blunders, or faults.}
\item{Q13}{I laugh and joke a lot with my closest friends.}
\item{Q14}{My humorous outlook on life keeps me from getting overly upset or depressed about things.}
\item{Q15}{I do not like it when people use humor as a way of criticizing or putting someone down.}
\item{Q16}{I don't often say funny things to put myself down.}
\item{Q17}{I usually don't like to tell jokes or amuse people.}
\item{Q18}{If I'm by myself and I'm feeling unhappy, I make an effort to think of something funny to cheer myself up.}
\item{Q19}{Sometimes I think of something that is so funny that I can't stop myself from saying it, even if it is not appropriate for the situation.}
\item{Q20}{I often go overboard in putting myself down when I am making jokes or trying to be funny.}
\item{Q21}{I enjoy making people laugh.}
\item{Q22}{If I am feeling sad or upset, I usually lose my sense of humor.}
\item{Q23}{I never participate in laughing at others even if all my friends are doing it.}
\item{Q24}{When I am with friends or family, I often seem to be the one that other people make fun of or joke about.}
\item{Q25}{I don't often joke around with my friends.}
\item{Q26}{It is my experience that thinking about some amusing aspect of a situation is often a very effective way of coping with problems.}
\item{Q27}{If I don't like someone, I often use humor or teasing to put them down.}
\item{Q28}{If I am having problems or feeling unhappy, I often cover it up by joking around, so that even my closest friends don't know how I really feel.}
\item{Q29}{I usually can't think of witty things to say when I'm with other people.}
\item{Q30}{I don't need to be with other people to feel amused - I can usually find things to laugh about even when I'm by myself.}
\item{Q31}{Even if something is really funny to me, I will not laugh or joke about it if someone will be offended.}
\item{Q32}{Letting others laugh at me is my way of keeping my friends and family in good spirits.}
}}
\source{
\url{http://personality-testing.info/_rawdata/HSQ.zip}
}
\usage{
HSQ
}
\description{
The Humor Styles Questionnaire (HSQ) data set comes from an interactive online version of the Humor Styles Questionnaire (Martin, Puhlik-Doris, Larsen, Gray, & Weir, 2003).
}
\details{
Responses for negatively worded statements (Q1, Q7, Q9, Q15, Q16, Q17, Q22, Q23, Q25, Q29, Q31) have been reverse-coded (i.e., 1=Very often or always true, 2=Often true, 3= Sometimes true, 4=Rarely true, 5=Never or very rarely true) in order to have the same direction (i.e., high values indicate more humorous personality) in all items.
}
\references{
Martin, R. A., Puhlik-Doris, P., Larsen, G., Gray, J., & Weir, K. (2003). Individual differences in uses of humor and their relation to psychological well-being: Development of the Humor Styles Questionnaire. Journal of Research in Personality, 37, 48-75.
}
\keyword{datasets}

