% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idi.R
\name{idi}
\alias{idi}
\title{Item Discrimation Index}
\usage{
idi(data, item, use = "complete.obs", na.rm = T)
}
\arguments{
\item{data}{data frame}

\item{item}{An item (as a vector)}

\item{use}{How should missing data be handled? See the \code{cor} function. Default is "complete.obs"}

\item{na.rm}{logical. Should missing values be removed? Defualt is TRUE.}
}
\value{
The item discrimination index
}
\description{
Calculates the item discrimation index
}
\examples{
idi(SAPA, SAPA$reason.4)
}
