% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iri.R
\name{iri}
\alias{iri}
\title{Item-Reliability Index}
\usage{
iri(data, use = "complete.obs", na.rm = T)
}
\arguments{
\item{data}{data frame}

\item{use}{How should missing data be handled? See the \code{cor} function. Default is "complete.obs"}

\item{na.rm}{logical. Should missing values be removed? Defualt is TRUE.}
}
\value{
The item-reliability index
}
\description{
Calculates the item-reliability index
}
\examples{
iri(SAPA)
}
