% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irtlr_summary.R
\name{irtlr_summary}
\alias{irtlr_summary}
\title{Summary of the IRT-LR DIF test from the mirt package}
\usage{
irtlr_summary(x, alpha = 0.05)
}
\arguments{
\item{x}{An object returned from the \code{mirt::DIF} function in the mirt package}

\item{alpha}{The alpha level of significance (default=0.05)}
}
\description{
This function creates a brief summary table for the \code{mirt::DIF} function in the mirt package. The printed output includes the flagged items, chi-square statistics, degrees of freedom, and corresponding p-values.
}
