% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irvi.R
\name{irvi}
\alias{irvi}
\title{Item-Reliability and Item-Validity Index}
\usage{
irvi(data, use = "complete.obs", na.rm = T)
}
\arguments{
\item{data}{data frame}

\item{use}{How should missing data be handled? See the \code{cor} function. Default is "complete.obs"}

\item{na.rm}{logical. Should missing values be removed? Defualt is TRUE.}
}
\value{
Item-Reliability and Item-Validity Index
}
\description{
Calculates the item-reliability index or the item-validity index
}
\examples{
cvr(20, 17)
}
