% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemperson.map.R
\name{itemperson.map}
\alias{itemperson.map}
\title{Item-Person Map for the Rasch Model and Its Variants}
\usage{
itemperson.map(x, y, k = 8, n = 10, ...)
}
\arguments{
\item{x}{Item difficulty paramaters for the items}

\item{y}{Latent trait (ability) estimates for persons}

\item{k}{Number of classes when grouping items and persons}

\item{n}{Number of persons represented by each X in the item-person map}

\item{...}{Additional elements to pass to the plot function}
}
\description{
Draws an item-person map for the Rasch Model and its variants using item difficulty parameters and latent trait estimates.
}
\examples{
## Item-person map for an exam with 15 items and 500 examinees
set.seed(355)
difficulty <- rnorm(15)
names(difficulty) <- 1:15
ability <- rnorm(500)
itemperson.map(difficulty,ability,main="Item-Person Map", n=20)

}

