% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivi.R
\name{ivi}
\alias{ivi}
\title{Item-Validity Index}
\usage{
ivi(item, crit, use = "complete.obs", na.rm = T)
}
\arguments{
\item{item}{An item (as a vector)}

\item{crit}{A criterion (as a vector)}

\item{use}{How should missing data be handled? See the \code{cor} function. Default is "complete.obs"}

\item{na.rm}{logical. Should missing values be removed? Defualt is TRUE.}
}
\value{
The item-validity index
}
\description{
Calculates the item-validity index
}
\examples{
ivi(SAPA$reason.4, reason.16)
}
