% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattice_pa.R
\name{lattice_pa}
\alias{lattice_pa}
\title{Parallel analysis plot via lattice}
\usage{
lattice_pa(data, ...)
}
\arguments{
\item{data}{A data.frame with no missing data}
}
\description{
\code{lattice_pa} returns a parallel analysis based on a reduced correlation matrix using lattice
}
\examples{
library(hemp)
cognition <- subset(interest, select = vocab:analyrea)
lattice_pa(cognition)
}
