% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattice_scree.R
\name{lattice_scree}
\alias{lattice_scree}
\title{Scree plot via lattice}
\usage{
lattice_scree(data)
}
\arguments{
\item{x}{A correlation matrix}
}
\description{
\code{lattice_scree} returns a scree plot using the lattice package
}
\examples{
lattice_scree(interest)
}
