% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latticePA.R
\name{latticePA}
\alias{latticePA}
\title{Parallel analysis plot via lattice}
\usage{
latticePA(data, ...)
}
\arguments{
\item{data}{A data.frame with no missing data}
}
\description{
\code{latticePA} returns a parallel analysis based on a reduced correlation matrix using lattice
}
\examples{
library(REPM)
cognition <- subset(interest, select = vocab:analyrea)
latticePA(cognition)
}

