% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mimic}
\alias{mimic}
\title{Multidimensional MIMIC Data Set}
\format{A dichotomous response data set with 2000 examinees and 24 items.
\describe{The item names range from "item1" to "item24". All the items are dichotomously scored. There is no missing data.
}}
\source{
Lee, S., Bulut, O., & Suh, Y. (2016). Multidimensional extension of multiple indicators multiple causes models to detect DIF. Educational and Psychological Measurement. Advance online publication. doi:10.1177/0013164416651116
}
\usage{
mimic
}
\description{
This data set has been fabricated using the item parameters in Lee, Bulut, and Suh's (2016) simulation study about
the multidimensional extension of the MIMIC model. The test structure involves within-item (i.e., complex) multidimensionality
where item1 to item6, item13 to item21, item23, and item24 are related to the first dimension; item7 to item20,
and item22 to item24 are related to the second dimension.
}
\keyword{dataset}
