% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multiplechoice}
\alias{multiplechoice}
\title{Hypothetical Multiple-Choice Test}
\format{A data frame in wide-format with 496 examinees' responses to 29 items.  
\describe{The first column is the examinee ID and the remaining columns are the responses to the items. Each item has four response options: 1=A, 2=B; 3=C; and 4=D. The answer key is as follows: 
key=c(4,3,2,3,4,3,2,3,1,4,3,2,3,3,4,2,4,2,3,3,1,2,2,1,2,1,1,2,1)}}
\usage{
multiplechoice
}
\description{
This hypothetical multiple-choice test consists of 29 items administered to 496 examinees. The data were randomly generated to demonstrate various psychometric models (e.g., 2PL-NLM and 3PL-NLM).
}
\keyword{dataset}
