% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_miss.R
\name{num_miss}
\alias{num_miss}
\title{Missing observations by items}
\usage{
num_miss(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
The number of missing observations, \code{num_miss}, and the percent missing, \code{perc_miss}.
}
\description{
Calcuate number of missing observations per items
}
\examples{
data(SAPA)
num_miss(SAPA)
}
