% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gstudy.R
\name{print.gStudy}
\alias{print.gStudy}
\title{Prints the results from the G-study}
\usage{
\method{print}{gStudy}(x, ...)
}
\arguments{
\item{x}{object of class \code{gStudy} returned from the \code{gstudy} function}

\item{...}{additional objects of the same type.}
}
\description{
Prints  output from fitting the \code{gstudy} function.
}
\seealso{
\code{\link{gstudy}}
}

