% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quart_cut.R
\name{quart_cut}
\alias{quart_cut}
\title{Quartile cutter}
\usage{
quart_cut(x)
}
\arguments{
\item{x}{A continuous variable}
}
\value{
A 4 level ordinal variable
}
\description{
Converts a continuous variable into an ordinal variable using quartiles.
}
\examples{
quart_cut(wiscsem$info)
}
