% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsd.R
\name{rmsd}
\alias{rmsd}
\title{Calculate RMSD}
\usage{
rmsd(theta1, theta2)
}
\arguments{
\item{theta1}{Vector containing the estimated latent traits from the first estimation method.}

\item{theta2}{Vector containing the estimated latent traits from the second estimation method.}
}
\description{
Calculates the root mean squared deviation.
}
\examples{
data(SAPA)
twopl_mod <- "F = 1 - 16"
twopl_fit <- mirt(data = SAPA, model = twopl_mod,
                  itemtype = "2PL", SE = TRUE)
latent_mle <- fscores(twopl_fit, method = 'ML', full.scores = TRUE, full.scores.SE = TRUE)
latent_map <- fscores(twopl_fit, method = 'MAP', full.scores = TRUE, full.scores.SE = TRUE)
rmsd(latent_mle, latent_map)
}
