% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rse}
\alias{rse}
\title{The Rosenberg Self-Esteem Scale}
\format{A data frame with 1000 participants who responded to 10 rating scale items in an interactive version of the Rosenberg Self-Esteem Scale (Rosenberg, 1965). There are also additional demographic items about the participants:
\describe{
\item{Q1}{I feel that I am a person of worth, at least on an equal plane with others.}
\item{Q2}{I feel that I have a number of good qualities.}
\item{Q3}{All in all, I am inclined to feel that I am a failure.}
\item{Q4}{I am able to do things as well as most other people.}
\item{Q5}{I feel I do not have much to be proud of.}
\item{Q6}{I take a positive attitude toward myself.}
\item{Q7}{On the whole, I am satisfied with myself.}
\item{Q8}{I wish I could have more respect for myself.}
\item{Q9}{I certainly feel useless at times.}
\item{Q10}{At times, I think I am no good at all.}
\item{gender}{Chosen from a drop down list (1=male, 2=female, 3=other; 0=none was chosen)}
\item{age}{Entered as a free response. (0=response that could not be converted to integer)}
\item{source}{How the user came to the web page of the RSE scale (1=Front page of personality website, 2=Google search, 3=other)}
\item{country}{Inferred from technical information using MaxMind GeoLite}
\item{person}{Participant identifier}
}}
\source{
The The Rosenberg Self-Esteem Scale is available at \url{http://personality-testing.info/tests/RSE.php}.
}
\usage{
rse
}
\description{
The RSE data set was obtained via online with an interactive version of the Rosenberg Self-Esteem Scale (Rosenberg, 1965). Individuals were informed at the start of the test that their data would be saved. When they completed the scale, they were asked to confirm that the responses they had given were accurate and could be used for research, only those who confirmed are included in this dataset. A random sample of 1000 participants who completed all of the items in the scale were included in the RSE data set. All of the 10 rating scale items were rated on a 4-point scale (i.e., 1=strongly disagree, 2=disagree, 3=agree, and 4=strongly agree). Items 3, 5, 8, 9 and 10 were reversed-coded in order to place all the items in the same direction. That is, higher scores indicate higher self-esteem.
}
\references{
Rosenberg, M. (1965). Society and the adolescent self-image. Princeton, NJ: Princeton University Press.
}
\keyword{dataset}
