% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_half.R
\name{split_half}
\alias{split_half}
\title{Split half reliability}
\usage{
split_half(data, type = "alternate")
}
\arguments{
\item{data}{A data frame}

\item{type}{Either alternate or random. Default is alternate.}
}
\value{
The split half reliability is caculated using every other item, if \code{type = "alternate"} is specified, or by splitting the instrument in half at random, if \code{type = "random"} is specifed.
}
\description{
Calculates split-half reliability either by selecting every other item or by randomly split the items.
}
\examples{
data(SAPA)
split_half(SAPA, type = "alternate")
split_half(SAPA, type = "random")
}
