% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.irtlr.R
\name{summary.irtlr}
\alias{summary.irtlr}
\title{Summary of the IRT-LR DIF test from the mirt package}
\usage{
\method{summary}{irtlr}(x, alpha = 0.05)
}
\arguments{
\item{x}{An object returned from the \code{mirt::DIF} function in the mirt package}

\item{alpha}{The alpha level of significance (default=0.05)}
}
\description{
This function creates a brief summary table for the \code{mirt::DIF} function in the mirt package. The printed output includes the flagged items, chi-square statistics, degrees of freedom, and corresponding p-values.
}
