% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swtable11}
\alias{swtable11}
\title{Item Scores on Eight Items from the CTBS Science Achievement Test}
\format{A data frame with 160 rows and 3 variables:
\describe{
  \item{person}{The person identifier}
  \item{score}{Whether a person got the item correct (1) or incorrect (0)}
  \item{item}{The item. There were 8 different items: i1 though i8.}
}}
\source{
Shavelson & Webb (1991). Generalizability Theory: A Primer
}
\usage{
swtable11
}
\description{
This is the data set presented in Table 1.1 in Shavelson & Webb (1991).
}
\keyword{dataset}

