% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teacher}
\alias{teacher}
\title{Item Scores on Eight Items from the CTBS Science Achievement Test}
\format{A data frame with 13600 rows and 6 variables:
\describe{
  \item{person}{The student identifier}
  \item{item}{Whether a person got the item correct (1) or incorrect (0)}
  \item{rater}{The four teachers in the study}
  \item{occasion}{Occassions 1 through 4}
  \item{rubric}{Whether a rubric was used or not.}
  \item{score}{Student score on the math problem. Ranged from 0 - 25}
}}
\source{
Kan, A., & Bulut, O. (2014). Crossed random-effect modeling: examining the effects of teacher experience and rubric use in performance assessments. Eurasian Journal of Educational Research, 57, 1-28. doi: dx.doi.org/10.14689/ejer.2014.57.4.
}
\usage{
teacher
}
\description{
The participants of this study were 50 eighth-grade students and 17 math teachers. The sample of teachers was selected to represent a range of experiences. Years of experience as a teacher ranged from one year to 26 years in the sample. Students, teachers, and performance tasks were fully crossed facets. All students in the sample responded to the same questions in a performance assessment. The responses of all the students were graded by each of the seventeen teachers.
}
\keyword{dataset}

