% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_length.R
\name{test_length}
\alias{test_length}
\title{New test length}
\usage{
test_length(data, r, r_type)
}
\arguments{
\item{data}{A data frame}

\item{r}{The desired reliability}

\item{r_type}{Either "alpha" for coefficient alpha,"split" for split-half reliability, or a numeric value given the current reliability calculated elsewhere. The "split" type assumes an even/odd split with the Spearman-Brown correction.}
}
\value{
The length of a test given the desired reliability
}
\description{
Calcuate the length of a test given a desired reliability.
}
\examples{
data(SAPA)
test_length(SAPA, r = .95, r_type = "split")
test_length(SAPA, r = .95, r_type = "alpha")
test_length(SAPA, r = .95, r_type = .90)
test_length(SAPA, r = .85, r_type = .90)
}
