% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threepl_demo.R
\name{threepl_demo}
\alias{threepl_demo}
\title{Three-parameter logistic IRT model Shiny demo}
\usage{
threepl_demo()
}
\description{
This code is a demonstration of Shiny and how Shiny can be used to explore the 3PL model.
}
\examples{
threepl_demo()
}
