% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_elements.R
\name{unique_elements}
\alias{unique_elements}
\title{Calculate number of unique elements in a covariance matrix}
\usage{
unique_elements(x)
}
\arguments{
\item{x}{A fitted lavaan object}
}
\description{
This function simply returns the number of unique elements in a covariance matrix. Recall that a correlation matrix is just a standardized covariance matrix.
}
\examples{
iq.cfa <- '
verb =~ info + comp + arith + simil + digit + vocab
perf =~ pictcomp + parang + block + object + coding
'
fit <- cfa(iq.cfa, data = wiscsem)
unique_elements(fit)

}

