% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiscsem}
\alias{wiscsem}
\title{WISC-R scores on 175 children}
\format{A data frame with 175 rows and 13 variables:
\describe{
  \item{info}{Information}
  \item{comp}{Comprehension}
  \item{arith}{Arithmetic knowledge}
  \item{simil}{Similarities}
  \item{vocab}{Vocabulary}
  \item{digit}{Digit span}
  \item{pictcomp}{Picture comprehension}
  \item{parang}{Picture arrangement}
  \item{block}{Block design}
  \item{object}{Object assembly}
  \item{coding}{Coding}
}}
\source{
\url{http://psych.colorado.edu/~carey/Courses/PSYC7291/ClassDataSets.htm}
}
\usage{
wiscsem
}
\description{
The WISC-R subscale data from Tabacknick & Fidell, 3rd ed "Using Multivariate Statistics". The suscale scores come for the second version of the Wechsler Intelligence Test for Children. The students were school-aged children previously diagnosed as learning-disabled. The data were collected between 1972 and 1984.
}
\keyword{dataset}

