% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{writing}
\alias{writing}
\title{Hypothetical writing prompt example}
\format{A data frame with 100 rows and 4 variables:
\describe{
  \item{student}{The student identifier}
  \item{prompt}{The writing prompt identifier. There were five prompts}
  \item{rater}{The rater identifier. There were two raters.}
  \item{score}{Student score on the writing prompt. Ranged from 4 - 100, with a maximum score of 100.}
}}
\source{
Desjardins, C. D. & Bulut, O. (2017). Using R for Educational and Psychological Measurement.
}
\usage{
writing
}
\description{
This is a hypothetical data set to demonstrate a two facet cross design for generalizability theory. The design is an S X W X R design, a student by writing prompt by rater design. See chapter 3 of "Using R for Educational and Psychological Measurement" for more details.
}
\examples{
library(lme4)
two.facet <- lmer(scores ~ (1 | students) + (1  | prompts) + (1 | raters) + (1 | students:prompts) +  (1 | students:raters) + (1 | prompts:raters), data = writing)
gstudy(two.facet)
}
\keyword{dataset}
