% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.plot.R
\name{design.plot}
\alias{design.plot}
\title{Plots various D studies for one and two-facet design}
\usage{
design.plot(x, unit, facets, g.coef = T)
}
\arguments{
\item{x}{object of class \code{mer} returned from the \code{lme4} function}

\item{unit}{the name of the unit of measurement}

\item{facets}{A list of length 1 (one-facet) or 2 (two-facet)}

\item{g.coef}{Plot the generalizability coefficent? If false, plot the dependability coefficientå}
}
\description{
Plots multiple d studies for one or two-facet design.
}
\seealso{
\code{\link{dstudy}}
}

