% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstudy.R
\name{dstudy}
\alias{dstudy}
\title{Conduct a dependability study}
\usage{
dstudy(x, n, unit)
}
\arguments{
\item{x}{A gstudy object}

\item{n}{A vector with facets named}

\item{unit}{The label for the unit of measurement}
}
\description{
Generates the variance components and the relative magnitude (percent of total variance) of each variance component.
}
\examples{
## One facet design
one.facet <- lmer(Score ~ (1 | Participants) + (1 | Items), data = efData)
one.facet.gstudy <- gstudy(one.facet)
dstudy(one.facet.gstudy, n = c("Items" = 4), unit = "Participants")

## Two facet design
two.facet <- lmer(scores ~ (1 | students) + (1  | prompts) + (1 | raters) + (1 | students:prompts) +  (1 | students:raters) + (1 | prompts:raters), data = writing)
two.facet.gstudy <- gstudy(two.facet)
dstudy(two.facet.gstudy, n = c("raters" = 2, "prompts" = 5), unit = "students")

## Two facet fixed design wtih prompt
gstudy.fixed.prompt <- gstudy(two.facet, fixed = "prompts")
dstudy(gstudy.fixed.prompt, n = c("raters" = 2), unit = "students")
}

