% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstudy.R
\name{dstudy}
\alias{dstudy}
\title{Conduct a dependability study}
\usage{
dstudy(x, n, unit)
}
\arguments{
\item{x}{An lmer fitted object}

\item{n}{A vector with facets named}

\item{unit}{The label for the unit of measurement}
}
\description{
Generates the variance components and the relative magnitude (percent of total variance) of each variance component.
}
\examples{
mod <- lmer(score ~ (1 | participant) + (1 | item), data = efData)
dstudy(mod)
}

