% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{efData}
\alias{efData}
\title{Hypothetical Executive Functioning Assessment}
\format{A data frame with 30 participants and 10 items.}
\usage{
efData
}
\description{
The executive function assessement consists of 10 items measured on 30 student participants. The data were randomly generated to demonstrate various psychometric models (e.g., CTT, a one-facet G-study)
}
\keyword{dataset}

